Rename dashboard page templates to pattern slugs
The template naming convention asks a slug to name the reusable UI
pattern, not the data or the task it happens to ship, so that a
neighbouring request matches it. Five dashboards named the fixture:

  dashboard-data                -> dashboard-comparison
  dashboard-executive-summary   -> dashboard-scorecard
  dashboard-portfolio           -> dashboard-composition
  dashboard-project-status      -> dashboard-progress
  dashboard-service-monitoring  -> dashboard-alert-rail

`dashboard` and `dashboard-cohort-funnel` already comply and are
untouched: the bare head is the family default, as with table/ and
login/, and funnel and cohort are chart archetypes rather than domains.

Two categories move with their slugs. dashboard-comparison takes
'Dashboard - Comparison' -- it previously shared 'Dashboard - Analytics'
verbatim with the dashboard template, so neither owned its strongest
keyword -- and dashboard-scorecard takes 'Dashboard - Scorecard'. Both
join the TemplateCategory union; the superseded values stay reserved.

Domain vocabulary is untouched in every description, which is where
retrieval reads it from. Also fixes "Executive Summary Dashoard".

Breaking: template resolution is exact-match on the directory name with
no alias mechanism, so the old slugs stop resolving.
