+++ packages/cli/assets/templates/pages/dashboard-alert-rail/page.tsx
