+++ packages/cli/assets/templates/pages/dashboard-alert-rail/template.doc.mjs
