+++ packages/cli/assets/templates/pages/dashboard-comparison/page.tsx
