+++ packages/cli/assets/templates/pages/dashboard-comparison/template.doc.mjs
@@ -7,6 +7,6 @@ export const doc = {
   displayName: 'Data Dashboard',
   description: 'Comparison-heavy analytics behind a persistent filter bar: tiles that each carry an inline sparkline and several period-over-period deltas, then segmented breakdowns underneath. Every number is shown against another number. Dashboard, metrics, stats, analytics, trends, or reporting.',
   isReady: false,
-  category: 'Dashboard - Analytics',
+  category: 'Dashboard - Comparison',
   isHiddenFromOverview: true,
 };
