+++ packages/cli/assets/templates/pages/dashboard-composition/page.tsx
