+++ packages/cli/assets/templates/pages/dashboard-composition/template.doc.mjs
