+++ packages/cli/assets/templates/pages/dashboard-progress/page.tsx
