+++ packages/cli/assets/templates/pages/dashboard-progress/template.doc.mjs
