+++ packages/cli/assets/templates/pages/dashboard-scorecard/page.tsx
