+++ packages/cli/assets/templates/pages/dashboard-scorecard/template.doc.mjs
@@ -3,10 +3,10 @@
 /** @type {import('@astryxdesign/cli/authoring').TemplateDoc} */
 export const doc = {
   type: 'page',
-  name: 'Executive Summary Dashoard',
+  name: 'Executive Summary Dashboard',
   displayName: 'Executive Summary Dashboard',
   description: 'Report-shaped analytics with a prose rail beside the charts: a scorecard of headline numbers with period deltas and status coloring, goal-attainment bars, a trend grid, and a written commentary column that folds beneath the content when narrow. Executive summary, business review, readout, or reporting.',
   isReady: true,
-  category: 'Dashboard - Executive Summary',
+  category: 'Dashboard - Scorecard',
   isHiddenFromOverview: true,
 };
