+++ packages/cli/authoring/doctypes/template/type.ts
@@ -86,9 +86,11 @@ export interface PageTemplateDoc extends BaseTemplateDoc {
 export type TemplateCategory =
   // Dashboard
   | 'Dashboard - Analytics'
+  | 'Dashboard - Comparison'
   | 'Dashboard - KPI Summary'
   | 'Dashboard - Monitoring'
   | 'Dashboard - Executive Summary'
+  | 'Dashboard - Scorecard'
   | 'Dashboard - Widget Grid'
   | 'Dashboard - Split'
   | 'Dashboard - Tabbed'
